+++ packages/docs/.storybook/preview.ts
@@ -1,5 +1,6 @@
 import type { Preview } from "@storybook/react";
 import { themes } from '@storybook/theming';
+import '../../docs/main.css';
 
 const preview: Preview = {
   parameters: {