+++ packages/react/src/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 export * from './components/Box'
 export * from './components/Text'
 export * from './components/Heading'
@@ -7,5 +9,3 @@ export * from './components/TextInput'
 export * from './components/TextArea'
 export * from './components/CheckBox'
 export * from './components/MultiStep'
-
-export * from './styles'