+++ packages/tokens/src/colors.ts
@@ -1,19 +1,23 @@
-export const colors = <const>{
+export const colors = {
   white: '#fff',
   black: '#000',
+  transparent: 'transparent',
 
-  gray100: '#E1E1E6',
-  gray200: '#A9A9B2',
-  gray300: '#C4C4CC',
-  gray400: '#7C7C8A',
-  gray500: '#505059',
-  gray600: '#323238',
-  gray700: '#29292E',
-  gray800: '#202024',
-  gray900: '#121214',
-
-  green300: '#00B37E',
-  green500: '#00875F',
-  green700: '#015F43',
-  green900: '#00291D',
+  gray: {
+    100: '#E1E1E6',
+    200: '#A9A9B2',
+    300: '#C4C4CC',
+    400: '#7C7C8A',
+    500: '#505059',
+    600: '#323238',
+    700: '#29292E',
+    800: '#202024',
+    900: '#121214',
+  },
+  green: {
+    300: '#00B37E',
+    500: '#00875F',
+    700: '#015F43',
+    900: '#00291D',
+  },
 }