+++ packages/tokens/src/font-sizes.ts
@@ -1,4 +1,4 @@
-export const fontSizes = {
+export const fontSize = {
   xxs: '0.625rem',
   xs: '0.75rem',
   sm: '0.875rem',