+++ packages/tokens/src/line-heights.ts
@@ -1,4 +1,4 @@
-export const lineHeights = {
+export const lineHeight = {
   shorter: '125%',
   short: '140%',
   base: '160%',